finish custom view for laying out cards
